+++ app/src/main/java/com/kongjak/koreatechboard/constraint/FCMTopics.kt
@@ -1,4 +1,4 @@
-package com.kongjak.koreatechboard.util
+package com.kongjak.koreatechboard.constraint
 
 const val FCM_TOPIC_SCHOOL = "school"
 const val FCM_TOPIC_DORM = "dorm"
