+++ app/src/main/java/com/kongjak/koreatechboard/service/FCMService.kt
@@ -1,11 +1,10 @@
-package com.kongjak.koreatechboard.util
+package com.kongjak.koreatechboard.service
 
 import android.app.NotificationChannel
 import android.app.NotificationManager
 import android.app.PendingIntent
 import android.content.Intent
 import android.os.Build
-import android.util.Log
 import androidx.core.app.NotificationCompat
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
