+++ app/src/main/java/com/kongjak/koreatechboard/ui/main/settings/SettingsViewModel.kt
@@ -21,8 +21,8 @@ import com.kongjak.koreatechboard.domain.usecase.SetSchoolNoticeSubscribe
 import com.kongjak.koreatechboard.domain.usecase.SetShowArticleNumberUseCase
 import com.kongjak.koreatechboard.domain.usecase.SetUserDepartmentUseCase
 import com.kongjak.koreatechboard.ui.base.BaseViewModel
-import com.kongjak.koreatechboard.util.FCM_TOPIC_DORM
-import com.kongjak.koreatechboard.util.FCM_TOPIC_SCHOOL
+import com.kongjak.koreatechboard.constraint.FCM_TOPIC_DORM
+import com.kongjak.koreatechboard.constraint.FCM_TOPIC_SCHOOL
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
